The project has officially begun. I changed the backend folders and files that were from the old project. There are a lot of thing that were done today, here is a list of them:
- Basics dependencies were installed, such as:
      "bcrypt": "^5.1.0",
      "cors": "^2.8.5",
      "dotenv": "^16.3.1",
      "express": "^4.18.2",
      "express-session": "^1.17.3",
      "jsonwebtoken": "^9.0.0",
      "mongodb": "^5.6.0",
      "mongoose": "^7.3.1",
      "passport": "^0.6.0",
      "passport-jwt": "^4.0.1",
      "passport-local": "^1.0.0"
And nodemon was added so we can run the command npm start and there is
no need to be restarting the project every time we make a change.
DISCLAIMER: there are some process.env invocation, a .env file is
needed to run the project (it doesn't matter the values of the vars);
this variables are: PORT, SESSION_SECRET and JWT_SECRET.

The skeleton was also added, having now in the root (Backend) directory
an index.js (the entry_point/main) a tests folder (empty), a config folder, auth folder, routes folder, models, foulder and controllers folder.

CONFIG: In the config folder we have our db made with MongoDB (mongoose) that
is currently using the default url of mongo. We also have a middleware
file where we have all our middlewares such as jwt where we verify the
headear for a "secret_token" parameters (there is where this secret_token
should be found). This jwt middleware is not tested at all yet.

AUTH: Here we only have out passport configuration where we have our
own signup and login authentication (local) and the jwt strategy. The
authentication is not properly tested yet because but we can create a
user and login in succesfully using the authentication method.

ROUTES: Here we added the index and the user routes where we have the
login and signup with passport auth, there is also a passport jwt auth
but it was not tested at all.

MODELS: In this we'll finde the only model at the moment, the User Schema.
There we can se that there are also two functions, a pre save function that
hashes the password before saving the instance, and the isValidPassword that
we use when a login is attempted.

CONTROLLERS: This folder has only one file that has "garbage" in it. It does
nothing at all at the moment but we will use it later on.

Finally, from now one, there will be more commit with fewer changes, or at
least that is what I will try to do.

Hope this clarifies a little bit what was done in this commit.
Ivan.