In the db file we only added the usage of the .env vars. The README.md now has the .env vars requiered for this prject's stage explained, also there the build and run explanation was added.
Until this commit, we have two models (user and admin) that use passport for authentication and jwt for verification (with passport), there are only three routes for each model, signup, login and get by id. We have urlencoded and bodyParser middlewares active, and cors and verifyAuthentication middlewares inactive. Test are not implemented yet but we tested the endpoint, auth and token with postman and they work fine.